requirements.txt
selenium
Appium-Python-Client
configparser
requests
beautifulsoup4
prettytable